Bulk Actions for Detections
Fixed a few styles to line up form elements with text.

Global constants are no longer accessible from the template. Moment is now specifically assigned to the component so it's use as a method is available.

Refactored how bulk actions cycle through the indeterminate state. The tri-state `selectAllState` was being perceived as 2x true states and 1x false state so indeterminatism is now represented by a second bool and the logic has been shuffled a little to achieve the old behavior.