Jobs/Job
Refactor the Jobs and Job pages for Vue 3.

V-data-tables with `:sort-by="..."` won't update the sort on change, added a few `@update:sort-by` clauses to important tables to remedy it.

Most of the JS changes relate to letting the v-data-table manage how rows expand. Although if you want to expand rows from outside the table (i.e. from buttons on the toolbar) then we still hook into `:expanded` and `@update:expanded`.