Use Server Context for ConfigStore Sync Actions
When a user updates a detection and has the proper detection perms, that detection should sync. For suricata, we were using the requester's context for all sync operations including reading/writing the necessary settings. This is an implementation detail that the requester should not be held to, so ConfigStore operations now use the server's context and the handler checks for permissions earlier.