Custom Functional TreeView
VTreeview is in Vuetify Labs still. Labs components are not prod-ready and when I tried to use them anyway the tab would crash when trying to display advanced settings. 3rd party solutions weren't packaged in an easy to use way for our Global Vue, non-typescript setup. So I created a new component from scratch.

Having advanced mode turned on WILL slow things down but the tab doesn't crash anymore. This solution is only to allow us to move to Vue 3 without the Vuetify Treeview. Once the official Treeview is out of Labs, we should probably use it over this custom component.

There's plenty of room for optimization, but this is a good square 1 to start from.