TouchUps for Alerts Cypress Tests
I missed the "switch to hunt" button not being present at the end of the search line. Just a matter of changing slots to Vue 3's expectations.

Removed instances of Vue.delete() that I didn't find the first time around. Due to Vue's improved reactivity a splice is all that's needed.

Removed an errant `debugger` statement that never should have been checked in.