Tweak Detections for Cypress Tests
Figured out how to custom sort time values. Needed to figure this out for a detections test that comments on a detection and checks that the comment is the newest entry in the history table. Updated other instances of dateAware sorting. Updated jest tests. All previous tests are passing.

Implemented headers correctly for detections history. Now it shows the sort icon and has other important styles.

Fixed snackbars not having x icons to close them.