Override Headers
Updated override headers to match between Elastalert and Suricata. I noticed we were specifying a column as "IP/Var" but was only showing IP so added a fallback property so these columns are more useful for other override types. The new Track column is split with regex so Modify can show it's regex and value.

Still need to think about what to do for Threshold overrides as they leave the ip/var column blank.