Detection Override Notes/Improvements
Override Notes are implemented. A hardcoded length value was used over a configured value so users don't enlarge the value and use these notes instead of comments.

These notes do not appear in a detection's history or create new audit entries when modified.

When searching detections, there's a new column indicating how many overrides exist for that detection.

The "Tuning" Tab on the detection page is now the "Tuning (#)" tab and contains a count of overrides on the detection.

Note that both of the above counts are total overrides, not enabled overrides.

A few HTML strings were updated to use i18n instead.

isFieldValid was refactored to check for an error state by checking the classes on the component. Vue 3 does not seem to offer a programmatic way to validate individual components.