Properly Supporting ClassList
ClassList is a special type called a DOMTokenList that has a contains method. Using `indexOf` sets off an infinite loop in Vue so the test needs to use a special object to emulate the parts of the DOMTokenList interface that we use. Also simplified how we reference classList.

If the ref is found, but the classList is not, the element should not be considered valid. Updated the function to check for this.