Matched columns to needs
Sigma doesn't show anything in the 2 multipurpose columns, removed them. Not much to show for a customFilter field that should contain multiple lines of YAML so leaving it out.

Refactored altValue to altValues, an array. The new pickValue function is used to select the value to show in the column by showing the first field that's truthy. Added test. Will get cypress test.