Fix UI Behavior Around Users
Expanding multiple users at once was resulting in the edit forms under each expanded row containing the values of the last row expanded. Reverted the behavior back to how it worked in Vue 2 where only 1 row could be expanded at once.

Similar to the above, if you expanded a row for an existing user and clicked the Create User button, then the data in the expanded row would be cleared out. Now when you click the Create User button, we close any expanded rows.

Added a test.