Quick Action Behavior and Popup Positions
Quick Action menu now has a specified width. In Vue 2, an unspecified width let it size itself to it's content. In Vue 3, we're using the target property and the unspecified width was resulting in the menu being as wide as the target. Now it looks good for all targets.

In order to differentiate between a click and drag event vs a simple click event when opening the Quick Action menu, we first check the window's selection to see if a Range is selected. If so, we don't show the menu, otherwise it appears normally.

Updated the banner tip to use Vue 3's location property instead of Vue 2's directives for positioning.