Logger on Context
Instead of creating arbitrary loggers or logging using the namespace name, the basepreprocessor has been updated to add a logger to every request's context that comes through. As we pass the context around, the logger goes with it. The logger gets the requestId field added to it so, similar to intCheckId and syncId, you can follow a request around by all the things logged with the same id.

Because we're using log.FromContext, if the context does not have a logger, we'll fall back to the global logger. This means no change to any log statement that receives a non-request context such as in modules or background processes.

I've updated every reference I could find looking for functions that make log namespace calls that had a context passed into them.

Not only does this give us a traceable requestId to help with debugging, this allows us to mock the logger in unit tests.