VSwitch Colors, QuickAction Position, Chart Redrawing
All v-switches now have a global color specified ("primary") and a new css rule that turns the "thumb" of the button "primary" to match. All individual color properties for switches have been removed.

The Quick Action menu will now be located where the user clicked instead of attached to the control the user clicked. This is especially noticeable on large strings.

Charts now have a "key" field that must be edited for the chart to redraw. Vue 3 will re-render components when their props change, but not if a field on their props changes. In order for Vue to notice that the Basic Metrics charts have new data, we need to specify a :key and change it.