Adapt to V-Combobox Oddities
v-checkbox should continue to use @change instead of @update:model-value.

v-combobox has an undocumented @change event that's closes to what we want and I'm adding Save/Cancel buttons to round out a good user experience for all use cases. Also, for non-`multiple` v-comboboxes, respect Enter to stop editing and commit the changes.

Improved the styling of the buttons in the attachments and observables table rows. They're more consistent with the rest of the app.

Fixed a bug around boxes not showing up in some v-comboboxes.