Fix V-Btn with Href
V-Btn's render differently with a hyperlink on them. Various colors change to blue to appear more link-like. I've added a few new styles to help keep buttons looking consistent on the PCAP page.

Removed theme-icon from several icons on the PCAP page. It may not be a good fit for the app at all.