Detection Panel Component
Add a new panel to Alerts. Clicking an alert will lookup the detection by it's publicId and show a panel with the summary and tuning tabs from the detections page.

On the backend, the route to get a detection by publicId now adds AI Summary metadata to the detection before returning it.

Added the copyright text to the treeview component when adding it to the new DetectionPanel component.

Can currently enable/disable the detection, and CRUD tunings.