Fix Duplicate Hunt on Navigation to Hunt Page
In the Vue transition, I added a call to loadData at the end of initHunt because loading the page was not executing the default query and this change seemed to fix it. I didn't know it at the time, but it resulted in the first query when navigating to the hunt page being executed twice.

The actual root of the problem is that navigating to your exact, current route does not navigate. This results in $route not changing, which means the watch for it doesn't call loadData. This was discovered during the Vue transition but the fix was only applied to pushing a new state (hunt.js:367) and was not applied to when we replace the state. The same fix that was applied then has been applied more thoroughly and now there's only 1 hunt executed during page load.