Don't Specify Chart Width
Some charts were appearing narrower than they should. instead of specifying both a height and a width for the charts to fit themselves to, only specify the height. This should fix the old issue of charts erroneously growing to take up all available vertical space in a bad way, and instead should ensure the chart takes up all available horizontal space in a good way. The choice to specify the height on the group div allows for the height to be specified while not maximized, but when the maximized class is applied the max-height value is overwritten.

Also fixed some icons who's styles had them hardcoded as white. They disappeared in the white theme.