await populateUserDetails
populateUserDetails is async and not finishing before the UI renders. This is causing the `owner` field on comments to render as undefined and when updated asynchronously, the UI isn't reflecting the change. Despite the JS comment saying otherwise, awaiting here is usually not waiting on a request. We cache our user list so repeated calls take trivial amounts of time. Added `await` and removed the comment so all comment owners appear in the UI properly.