Use Refs for Performant Comments
Before a case looks up the owner or assignee for one of it's associations, pre-load the field with a ref. This provides a much faster way to add the owner asynchronously if we have to look it up.

Update populateUserDetails to support Refs by properly assigning the value to `.value` when the target field is a ref and directly to the target field when it's not a ref.

Updated an app test to pass in a ref and non-ref object to it's populateUserDetails test and test both outcomes accordingly. Updated test_common to include all the work necessary to support Vue.ref and Vue.isRef through re-implementing them.