Details Panel ++
Show the panel using a toggle. Defaults to true. When not visible, the info buttons disappear and the events take up the full page width. When visible, it now shows a little help text in the space until you select an event. Only appears on the Alerts page. The preference is saved in local storage.

The details panel now scrolls separately from the main events table.

Each row has an info icon that when clicked opens the detection in the details panel and changes the icon to thumbtack. Clicking the thumbtack does nothing.

The QuickActionMenu now remembers what subgroups were opened by the user and keeps them open until they refresh their session (i.e. it's not stored in local storage so refreshing the page results in all groups being closed by default again).

Used new technique in jest tests to cause any outstanding promises to finish before assertions are asserted. This exposed an issue with a setTimeout in the "relative query string" test. That was fixed by mocking the problem function and using the same technique to wait out the timeout.

Add annotations to the quick links in the details panel. The buttons also have less space around them leaving room for when we add more.

Fixed an issue in Cases where the description wasn't being rendered. Turns out we should use `v-html` over `:inner-html.prop` so I updated the only other reference to the inner-html: the login page.

Added some checks to replaceActionVar. Test data includes non-latin characters that don't convert to base64. To fix this with the least amount of ripple, I added a check to make sure that the target actually uses that {var} before base64 encoding so it doesn't throw an error trying to generate data for a variable that's not even used.

ToDo: Look for a better way to maintain open group history for the quick action menu. Opening the menu currently clears the open array and I'm not sure why. There's a hack to reset the array after the menu becomes visible.