Proper Maximizing of Charts
Maximized charts get the `maximized` class which sets `position: absolute` among other things. For proper positioning, no parent of the chart-container should be `position: relative`. Removed this property from the `groupBy` class.

Tested a very large table and the table body is scrollable. There was previously concern that the page hiding it's scrollbar when a chart is maximized would hide some of these details but that is not the case. Tested with the query `event.dataset:"system.syslog" | groupby @timestamp`