Test More Detection Handlers
Added go:generate line to mock DetectionEngine.

To prevent circular imports, handler functions need to be made public for tests.

Added tests for GetDetection and GetByPublicID providing 100% coverage of their handlers. The CreateDetection test is in-progress.

Added EntryMatcher as a way to test logs for particular levels, fields, and messages. The plan is the test creates a parallel array with conditions on each EntryMatcher that test their corresponding log entry. A matcher is stateless and can be reused.

Added a few helper functions.