Implement more Detection Handler Tests
While writing tests for CreateDetection, it was discovered that we call ApplyFilters twice and this was causing a logic error because we only tracked if the status changed the second time but the first time is when the change will take place. This prevented CreateDetection from ever indicating that the created detection was modified.

Built a better way to test responses with a type switch and updated existing tests to use it.

TODO: Evaluate if UpdateDetection should call ExtractDetails before PrepareToSave (which calls ExtractDetails again).