Fixed Issues Found While Adding Tests
Calling ExtractDetails directly before PrepareForSave is unnecessary as the first thing PrepareForSave does is call ExtractDetails.

UpdateDetection cannot return some of the errors that it checked for. Those errors were caught by PrepareForSave. Simplified the list of errors checked.

When a Detection is disabled because it failed to sync while enabled, we failed to call MergeAuxData. Refactored to fix this.

More tests for UpdateDetection.