Fix "Create Case" Error
When creating a case, we 1) make a call to create a case, 2) $router.replace to update the url with the new case ID, 3) use the route params to load the case. Step 2 is async so step 3 used the old ID param of `create`. Adding an await on step 2 ensures that the new route params are used in step 3 and we don't attempt to retrieve a case with a bad ID.