Update Selection
When clicking on a config value, we hide the text-area and invisibly swap out to a different text-area. It's almost invisible except that the focus isn't maintained: the user focused the old element but not the new one.

Simply adding a call to .focus() puts the cursor in a different place than where the user clicked. This isn't how text controls work and can break a user's flow.

The selection cursor isn't updated to the click location in the focus event handler so it was refactored to be a click handler. In the edit event that fires, the scroll position and selection range are taken from the old input and on next tick applied to the new input. This makes the control swapping truly invisible.

You still have to focus the text-area before the resize handle will be available though.