Use @focus over @click
Tabbing to the control wasn't giving it focus. Reverted the event back to focus and then wrapped the event body in a `setTimeout([event body], 1)` so it executes after the click event sets the selection range.

Now it works for node specific values as well by adapting the selector to the nodeId parameter.

Tabbing to the control puts the cursor at the end.

Clicking on the control puts the cursor where the user clicks.

Highlighting text in the control highlights the same text.

Methodology used: We setTimeout(..., 1) to let the click event finish so the cursor location is accurate. After that, we change our form variables normally. Then we collect scroll and caret information. We wait one tick for the form changes we just made to propogate through the UI and then we set the scroll and caret information on the new text-area.