V-Btn-Toggle Changes
In Vuetify 2, a v-btn-toggle updated it's model before executing it's childrens' events, in Vuetify 3 that's swapped to better align with Vue's reactivity rules. That means captureLayoutAsStream on the PCAP page is seeing an out-of-date model when it references packetOptions causing the button to act exactly backwards from it's indicated state. Waiting for nextTick allows the model to be updated and the correct state to be observed.