Replace Virtual Table
v-data-table-virtual says on the tin that it "relies on all data being available locally but unlike the standard data-table it uses virtualization to only render a small portion of the rows." This makes it not a good fit for use showing alert detail rows in the ungrouped Alerts table. We're only rendering a small portion of rows (a little over half) upon expanding a row. The behavior appears glitchy rather than intended when you expand a second row and the first row fills out it's missing rows.

Now the table is the non-virtual variant with items-per-page set to -1 so we see every row as expected.