The Rest of the Handlers
Reorganized a little. Repeat Log Entry Matchers were moved to package level variables because they're stateless and can be reused.

Added a default case in the Response object type check. This will help against a test appearing to pass if a test expects a type that the switch didn't account for.