Prevent Second Popup When Disabling Detection
When disabling a detection, we want to offer a popup to the user asking "Do you also want to acknowledge all alerts for this detection?" This popup is controlled by the Detection Panel and is desired. However if the user clicks Yes, the panel calls out to the alerts page to do the actual acknowledging. At the page level, a new optional parameter was added to the ack function to help with a soon-to-be-implemented feature where every event can be escalated and this too will have a dialog. The detection panel was failing to provide the new optional parameter that would hide this second page-level popup and thus 2 confirmation popups were being displayed. Having the detection panel pass `true` for this optional parameter prevents the second popup and provides the user with a cleaner experience as we originally intended.

Updated test.