Right Number of Dialogs at the Right Times, Check Language and Engine
Restored the ack all dialog triggered by the Detection Panel on alerts. Updated tests.

When converting a rule from the New Detection page, the detection's engine isn't specified yet. Refactored a check to now allow the engine to be "elastalert" or the language to be "sigma" so the request can continue as it should. Added a new test just for this.