Fix Grid Table Border and OIDC Alert Line Wrapping
The grid table had the `show-expand` property but since we're handling expansion ourselves, we don't need the extra table cell. Removed.

V-Alerts attempt to store their text on one line and hide any overflow. Added a class to ensure the text wraps instead for a v-alert on the user's page with a large amount of text.