Attach Many Events to Case, Detections Panel Escalate
Refactored the Case createEvent endpoint to createEvents, it now searches for events and attaches them to the case individually instead of as an aggregation.

Restored the Escalate button and it's behavior to the Detection Panel. Updated the body of the call to /case/events to correspond with the refactor.

TODO: Tests, update request body model for createEvents to stop using the modified RelatedEvent, update doc comments on modified endpoint, more tests