Limit Bulk Escalation
New client parameter for alerts to limit how many are escalated to a case at once.

If the group of detections to be escalated appears to be more than this limit, there's an extra warning on the confirmation dialog that includes a link to where the limit can be configured.

Under Case's Events tab, there's a hunt button to show these related events in Hunt.

After adding the related events to a case, extract observables.

While attempting to use Scrolling to search for related events, SortFields was added so the escalate limit would always grab the most recent. Eventually Search was used over Scroll because search allows for limiting the number of returned elements. I'm leaving sorting support in because it works.

When converting Events to RelatedEvents, copy the @timestamp field into the soc_timestamp field so the table of events can still show and sort by this value.

Update tests to both have an escalate limit set in the config and account for attaching observables to the case.

TODO: tests