Better "Hunt This Case"
huntCase now finds the earliest timestamp in the event list and sets the dateRange to span from that point to now to show every event.

filterVisibleFields now prepends common  "bundle" fields used in places like Related Events for a case. Events with these bundle fields will cause the headers to show them instead of the normal un-prefixed headers BUT still allow for eventFields to determine a helpful set of headers based on module and dataset.

AttachEventQuery was renamed to AttachEventCriteria to better match the names of other, similar Criteria types.

New jest tests