Faster/More Accurate Escalate
Pass Escalate and Acknowledge parameters to the endpoint so the backend search better aligns with the filters the user has set on the frontend.

Optimize how we process extracting Observables when attaching events to a case. Instead of iterating over every field, iterate only over the fields we're looking for.