Local Storage and Objects
sortBy is an array. When we save to localStorage, we need to save it as a string otherwise the value becomes `[object Object]` which breaks the headers of the table when the value is used onLoad. When storing, stringify. When reading from storage, parse. This does not effect other components as they break up how sort is stored by having "sortby" and "sortDesc" and not attempting to store objects.

Anybody currently affected by this bug can clear their cache or more specifically open DevTools > Application > Storage > Local storage > [soc url] and find the key "settings.jobs.sortBy" and delete it. Then refresh the page.