Custom Sort Severity
When constructing headers for the alerts table, add a special sort function to any columns for the value `event.severity_label`. This custom sort function orders the severity by intensity (low < medium < high) rather than alphabetically (high < low < medium).

Refactored updateGroupBySort and referenced in the UI. It's been reworked so it only saves when the first group is the one being modified.