Performance Profiling
This single file is only included in the build if the "debug" build flag is specified so by default the /debug/pprof endpoints are excluded from the build.

When included, this not only attaches the /debug/pprof endpoints, but this also provides a new middleware we can apply to endpoints in the app so that any calls to that endpoint will result in a cpu profile (and optional memory profile) being generated in your /tmp folder.

These tools are for the devs and should not make it into a production release. Although the performance impact of such an endpoint being released would be negligible, the security implications of somebody accessing the memory of a running service are too great to just add auth to the endpoint. This and it's accompanying Cypress test are designed to ensure there will not be any debug or pprof code in the binary coming out of CI/CD.