DateRange on PCAP Screen
I completed about 90% of the task before rereading the original story only to see I misinterpreted it. This isn't what was asked for but as it's so near completion I can't bring myself to delete it yet. Checking in so we can make a decision about it in the future.

Does have a bug: the relative time units drop down doesn't bind properly on page load, it shows the value when it should show the title of the matching item ("30" instead of "hours") but works properly.