Fix Always Passing Date Parameters to Backend
The old approach took advantage of the fact that undefined properties, when converted to JSON, are left out. The updated approach always specified the `beginDate` and `endDate` fields, if no value was entered by the user an empty string was used and this threw a 400 from the backend because they don't parse as a valid date. The new approach conditionally adds them if the user set their values.

Added a regression test.