Detection Panel Fixes
Add newOverrideValid to the model so we properly bind the property. This allows multiple overrides to be added without refreshing the page.

Add missing validation rule for notes.

Remove createOverrideTypeChange as we haven't been calling it (@change isn't the proper event and using the proper event creates an endless loop).

By wrapping the contents of the detection-col in a div with `position: relative;` we can ensure the loading spinner renders correctly at every scroll position.