Fix For Zoom Levels between %25 and %150
The page is 12 columns wide split into 4 and 8 with a divider in the middle. As zoom levels shrink, the divider's negative width isn't adjusted properly and this pushes the 8 columns to a new row. By decreasing the margin-left from -1px to -4px, zoom levels down to 25% will continue to show the 2 columns side by side without having a large impact on the un-zoomed appearance of the page. Zoom levels up to 150% look good and after that the view is no longer considered "Large" and the columns rearrange properly.