Multiline Comment Fix
In some situations, a /* multiline comment */ was being interpreted as quotes and breaking the import of 13 files. By clearing quotes when we exit comments, this issue is resolved. As a sister fix, if a strings variable opens a `{}` block of bytes to define, accept the newlines as part of the string definition and treat it like a multiline string.

Refactor a test to account for a multiline comment before a multiline string of bytes that should pass but previously wasn't.