Fix PCAP Date Range
With autoUpdateInput set to false, the daterangepicker wasn't updating the model. Once I fixed that, I saw we were sending the dates in the wrong format for Go's JSON parsing to successfully parse it as a time.Time so now we format it into ISO 8601.

Updated loadLocalSettings to restore the timeframe and not the 2 independent dates it used to be.