Conversion Adjustments
When we have the sigma cli convert a question's query, we need it to convert all the variable field names as well. The sigma cli pipeline has been updated to return a JSON object with the converted query and the new field names. Also, if the yaml document separator `---` is used to supply multiple queries to the sigma cli, multiple JSON objects will be returned. The playbook module has been updated to take advantage of these changes.

Improved logging.