Max Height on Input Messages
Persistent hints appear below an input and in some narrow window sizes would stack each letter of the text on top of the next displacing the rest of the page vertically. By limiting the max-height, we can control that behavior. The text will wrap until it's 3 lines tall and then it will start truncating text.