Pay Closer Attention to ModifiedByFilter
Updated applyStatusRegexes to only return true if the filter actually changed the status, not just if the filter would have changed the status. New test added to be sure.

Now, if a regex actually modifies the detection the detection will be saved. This means a user can add an enableRegex or a disableRegex and see the detections affected by them change to the correct status without developing a Rule Mismatch status for the engine.

Added a new sync test that normally would see 1 rule, see it's enabled and do nothing, except that the 1 rule matches the specified disableRegex so the test ensures that the detection is