Ungrouped Alerts
Load playbooks recursively from target directory.

Ensure Playbooks are loaded ASAP with firstRun sync.OnceFunc.

On alerts, in the ungrouped table, begin adding the tabs to the expanded row of an alert.

The playbooks are requested when the "Guided Analysis" tab is clicked. A loading circle displays in the tab body while the playbooks are fetched. Questions+Contexts are rendered in an accordion fashion.