Visual Cleanup and Bug Fixing
Visually cleaned up the progress so far because it'll be used as a template for other places (grouped alerts).

Only provide quick answers to the questions if the licenseState === 'active' using new $root.isPro.

Behave better around refreshing while rows are expanded. Probably still more to do on this front, but better.