New Variable Substitution Flow
Conversion was moved from SOC startup to upon request by the UI. Because of things like `startswith` in the query modifying variable values to use wildcards, we can't convert the sigma to OQL before variable substitution. So we no longer convert on startup, we deliver playbooks with unconverted sigma queries, the UI fills in the variables, the filled sigma is sent to the API to convert to OQL.

Changed field casing on the models returned by the API. Updated references in the UI.

New default playbook repo branch.