Fixes and Some Tests
Fixed the timezone issues in buildQuestionRange.

Removed some of the overreaching parts of getEventField for now. We'll tighten the playbook in places where the fields are wrong.

New way to retrieve a timestamp from an event that should work on any event given to it (an alert event that has original values under `event_data` or a regular event).

Todo: Move new strings to i18n, more tests.