Hash Repo Paths For Folder Names
We're running into collisions when people are adding sigma rules from repos with the same name. To fix this, we hash the entire repo URL and that becomes the folder name we clone into. On next sync, each engine will re-clone all the repos and delete the old non-hashed folders.

Update playbook to hash the folder name and follow the scheme when accessing playbooks in the git folder. This change isn't necessary for the detection engines as they use the paths returned by UpdateRepos that will be accurate.