Bidirectional Binding
Forgot to include bidirectional binding. The dropdown would default to 10 items per page. If the user changed it, the table would change behavior but we wouldn't notice the new page size to save it. Then when you refresh the page, it defaults of 10 again. We were already watching for the variable change to save it, but without bidirectional binding only the table saw the new value and not our component.