RulesetName
Add rulesetName to RuleRepo to allow the user to specify the name of the ruleset. The ruleset name is used as the folder name when a repo is pulled allowing the user to avoid collisions. If a rulesetName is not specified, the last folder of the repo will be used.

Updated how we clone the repos to use this new value as the folder name. Updated the engines to use this field as the ruleset of the detections found in that repo. Updated tests.