Add Permissions to Playbooks
I've added write and read permissions to the rbac files. I've added read permissions to the API. Write permissions will be used later.

I updated the Interrupt function to pass in a context and out an error so that it too can have permissions checked although it is currently unused.

Because modules are all started at approximately the same time, the playbooks module needs to wait for the rbac module to load it's files. New logic has been added to the playbook module to verify permissions before it begins. It will check for up to 15 seconds, after that it returns an error, writes to the log, and turns off the module. In a properly configured system, it may slow down the starting of the playbook module by a few hundred milliseconds at most.