Detections Playbook Tab
The /playbook/detection/{id} route respects a new query string parameter `raw`. If added on with a true value, the response will be in application/x-yaml. This is so the YAML is returned properly, avoiding a lot of problems around returning a YAML playbook in JSON and converting it back to YAML client side.

The detections page now has a Playbook tab where the source of all playbooks that apply to that detection are displayed in a read-only, syntax highlighted view.