Specify Origin in Config
We can no longer change the window.location.origin on the fly. The jest-environment-jsdom plugin now properly simulates window.location as a read-only object which is having side effects. To remedy, I'm setting the URL in the config so all window.location.origin checks are now non-localhost. This fixes a config test and impacts a hunt test. Refactored the hunt test to work with the new origin.

Removed old polyfill code from test_common.js.