Deterministic IDs and Detective Logging
Anytime a new detection is being added to ES, we now pre-generate it's ID by hashing the publicId. This should result in deterministic IDs that prevent multiple imports of community detections.

Adding logging that should only happen if a GetAllDetections call returns 0 detections.

Fix itemsperpage bug on detections screen. Needed v-model to bind in both directions.

Put syncLocalDetections on the detectionshandler.

Because we're now creating documents with pre-generated IDs, I need a way to call ElasticDetectionstore.save without using the ID as an indicator for if the audit object should be for a create or an update. Added a new context value that can override and specify that even though there's an ID, this document is being created.