Playbook LUTs
Playbooks are no longer stored in memory after being read in. Because of the large number of playbooks expected, the new approach reads all the playbooks once, organizes their IDs instead of the entire playbook, and keeps a reference of which IDs were found in which files on disk. When playbooks are requested, they're read from disk and returned.