State File But No Community
This adds logic and logging for a peculiar scenario in detections: a state file is present indicating that a previous sync has finished successfully during the same sync that a call to GetAllDetections returns 0 community detections. When this happens, the sync stops and is considered a failure. A user may delete the state file to allow the import to proceed. The detections page will show Sync Failed for the engine. The log will indicate "state file present but 0 community rules found" as the reason the sync failed.

If the Suricata engine finds itself in this failure state, it will NOT attempt migrations as the only current migration uses GetAllDetections and requires it to return properly. Again, deleting the state file will let the sync import and in Suricata's case the migrations will then run.