Refactoring/Testing
Show an error output when playbooks can't be fetched.

Clicking on an expand or collapse button before the playbook is fetched will fetch the playbook and then properly expand or collapse.

Collapse/Expand All buttons moved to tab.

Added some missing data-aid and aria-labels to some of the new additions.

Tests.