Remove Item-Value
Cannot specify the item-value of the alerts data table to `rule.uuid`. That works fine for "Group By Name, Module" but in any other view the `rule.uuid` of the alert may not be unique in the list of alerts. This is causing UI errors when a different grouping method is selected and the columns are then sorted, resulting in extra rows visible in the UI. Not specifying the `item-value` remedies the issue.