Give Alerts Unique IDs for the UI
We now generate a pseudo-ID for alerts. It only needs to be unique per object so it's the index at the time of retrieval. The IDs are then used for identifying individual rows when working with v-data-table. The new field has been exempted from existing functions that utilize all alert fields. These refactors make it possible to delete events and continue to use the page without refreshing but did require a bit of refactor. The result is a stabler state as the user moves around the Alerts page.